--- a/pkg/ccl/sqlccl/bench_test.go
+++ b/pkg/ccl/sqlccl/bench_test.go
@@ -15,6 +15,8 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/cockroachdb/cockroach/pkg/ccl/storageccl/engineccl"
+	"github.com/cockroachdb/cockroach/pkg/settings"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/tracing"
 )
@@ -133,4 +135,57 @@ func BenchmarkLoadSQL(b *testing.B) {
 		sqlDB.Exec(line)
 	}
 	b.StopTimer()
+}
+
+func runEmptyIncrementalBackup(b *testing.B) {
+	defer tracing.Disable()()
+
+	const numStatements = 100000
+
+	ctx, dir, _, sqlDB, cleanupFn := backupRestoreTestSetup(b, multiNode, 0)
+	defer cleanupFn()
+
+	restoreDir := filepath.Join(dir, "restore")
+	fullDir := filepath.Join(dir, "full")
+
+	ts := hlc.Timestamp{WallTime: hlc.UnixNano()}
+	if _, err := Load(
+		ctx, sqlDB.DB, bankStatementBuf(numStatements), "bench", restoreDir, ts, 0, restoreDir,
+	); err != nil {
+		b.Fatalf("%+v", err)
+	}
+	sqlDB.Exec(`DROP TABLE bench.bank`)
+	sqlDB.Exec(`RESTORE bench.* FROM $1`, restoreDir)
+
+	var unused string
+	var dataSize int64
+	sqlDB.QueryRow(`BACKUP DATABASE bench TO $1`, fullDir).Scan(
+		&unused, &unused, &unused, &dataSize,
+	)
+
+	// We intentionally don't write anything to the database between the full and
+	// incremental backup.
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		incrementalDir := filepath.Join(dir, fmt.Sprintf("incremental%d", i))
+		sqlDB.Exec(`BACKUP DATABASE bench TO $1 INCREMENTAL FROM $2`, incrementalDir, fullDir)
+	}
+	b.StopTimer()
+
+	// We report the number of bytes that incremental backup was able to
+	// *skip*--i.e., the number of bytes in the full backup.
+	b.SetBytes(int64(b.N) * dataSize)
+}
+
+func BenchmarkClusterEmptyIncrementalBackup(b *testing.B) {
+	b.Run("Normal", func(b *testing.B) {
+		defer settings.TestingSetBool(&engineccl.TimeBoundIteratorsEnabled, false)()
+		runEmptyIncrementalBackup(b)
+	})
+
+	b.Run("TimeBound", func(b *testing.B) {
+		defer settings.TestingSetBool(&engineccl.TimeBoundIteratorsEnabled, true)()
+		runEmptyIncrementalBackup(b)
+	})
 }