--- a/pkg/ccl/storageccl/engineccl/.gitignore
+++ b/pkg/ccl/storageccl/engineccl/.gitignore
@@ -1,0 +1,3 @@
+# Do not add environment-specific entries here (see the top-level .gitignore
+# for reasoning and alternatives).
+mvcc_data