--- a/pkg/ccl/storageccl/export.go
+++ b/pkg/ccl/storageccl/export.go
@@ -115,10 +115,9 @@ func evalExport(
 
 	// TODO(dan): Move all this iteration into cpp to avoid the cgo calls.
 	// TODO(dan): Consider checking ctx periodically during the MVCCIterate call.
-	iter := engineccl.NewMVCCIncrementalIterator(batch)
+	iter := engineccl.NewMVCCIncrementalIterator(batch, args.StartTime, h.Timestamp)
 	defer iter.Close()
-	iter.Reset(args.Key, args.EndKey, args.StartTime, h.Timestamp)
-	for ; iter.Valid(); iter.Next() {
+	for iter.Reset(args.Key, args.EndKey); iter.Valid(); iter.Next() {
 		if log.V(3) {
 			v := roachpb.Value{RawBytes: iter.UnsafeValue()}
 			log.Infof(ctx, "Export %s %s", iter.UnsafeKey(), v.PrettyPrint())