--- a/pkg/storage/engine/db.cc
+++ b/pkg/storage/engine/db.cc
@@ -61,7 +61,7 @@ struct DBEngine {
   virtual DBStatus ApplyBatchRepr(DBSlice repr, bool sync) = 0;
   virtual DBSlice BatchRepr() = 0;
   virtual DBStatus Get(DBKey key, DBString* value) = 0;
-  virtual DBIterator* NewIter(bool prefix) = 0;
+  virtual DBIterator* NewIter(rocksdb::ReadOptions*) = 0;
   virtual DBStatus GetStats(DBStatsResult* stats) = 0;
 
   DBSSTable* GetSSTables(int* n);
@@ -71,7 +71,6 @@ struct DBEngine {
 struct DBImpl : public DBEngine {
   std::unique_ptr<rocksdb::Env> memenv;
   std::unique_ptr<rocksdb::DB> rep_deleter;
-  rocksdb::ReadOptions const read_opts;
   std::shared_ptr<rocksdb::Cache> block_cache;
   std::shared_ptr<DBEventListener> event_listener;
 
@@ -102,14 +101,13 @@ struct DBImpl : public DBEngine {
   virtual DBStatus ApplyBatchRepr(DBSlice repr, bool sync);
   virtual DBSlice BatchRepr();
   virtual DBStatus Get(DBKey key, DBString* value);
-  virtual DBIterator* NewIter(bool prefix);
+  virtual DBIterator* NewIter(rocksdb::ReadOptions*);
   virtual DBStatus GetStats(DBStatsResult* stats);
 };
 
 struct DBBatch : public DBEngine {
   int updates;
   rocksdb::WriteBatchWithIndex batch;
-  rocksdb::ReadOptions const read_opts;
 
   DBBatch(DBEngine* db);
   virtual ~DBBatch() {
@@ -123,7 +121,7 @@ struct DBBatch : public DBEngine {
   virtual DBStatus ApplyBatchRepr(DBSlice repr, bool sync);
   virtual DBSlice BatchRepr();
   virtual DBStatus Get(DBKey key, DBString* value);
-  virtual DBIterator* NewIter(bool prefix);
+  virtual DBIterator* NewIter(rocksdb::ReadOptions*);
   virtual DBStatus GetStats(DBStatsResult* stats);
 };
 
@@ -143,19 +141,16 @@ struct DBWriteOnlyBatch : public DBEngine {
   virtual DBStatus ApplyBatchRepr(DBSlice repr, bool sync);
   virtual DBSlice BatchRepr();
   virtual DBStatus Get(DBKey key, DBString* value);
-  virtual DBIterator* NewIter(bool prefix);
+  virtual DBIterator* NewIter(rocksdb::ReadOptions*);
   virtual DBStatus GetStats(DBStatsResult* stats);
 };
 
 struct DBSnapshot : public DBEngine {
   const rocksdb::Snapshot* snapshot;
-  rocksdb::ReadOptions read_opts;
 
   DBSnapshot(DBEngine *db)
       : DBEngine(db->rep),
-        snapshot(db->rep->GetSnapshot()) {
-    read_opts.snapshot = snapshot;
-  }
+        snapshot(db->rep->GetSnapshot()) {}
   virtual ~DBSnapshot() {
     rep->ReleaseSnapshot(snapshot);
   }
@@ -168,7 +163,7 @@ struct DBSnapshot : public DBEngine {
   virtual DBStatus ApplyBatchRepr(DBSlice repr, bool sync);
   virtual DBSlice BatchRepr();
   virtual DBStatus Get(DBKey key, DBString* value);
-  virtual DBIterator* NewIter(bool prefix);
+  virtual DBIterator* NewIter(rocksdb::ReadOptions*);
   virtual DBStatus GetStats(DBStatsResult* stats);
 };
 
@@ -198,6 +193,20 @@ rocksdb::Slice ToSlice(DBString s) {
 
 const int kMVCCVersionTimestampSize = 12;
 
+void EncodeTimestamp(std::string& s, int64_t wall_time, int32_t logical) {
+  EncodeUint64(&s, uint64_t(wall_time));
+  if (logical != 0) {
+    EncodeUint32(&s, uint32_t(logical));
+  }
+}
+
+std::string EncodeTimestamp(DBTimestamp ts) {
+  std::string s;
+  s.reserve(kMVCCVersionTimestampSize);
+  EncodeTimestamp(s, ts.wall_time, ts.logical);
+  return s;
+}
+
 // MVCC keys are encoded as <key>[<wall_time>[<logical>]]<#timestamp-bytes>. A
 // custom RocksDB comparator (DBComparator) is used to maintain the desired
 // ordering as these keys do not sort lexicographically correctly.
@@ -210,12 +219,7 @@ std::string EncodeKey(DBKey k) {
     // Add a NUL prefix to the timestamp data. See DBPrefixExtractor.Transform
     // for more details.
     s.push_back(0);
-    EncodeUint64(&s, uint64_t(k.wall_time));
-    if (k.logical != 0) {
-      // TODO(peter): Use varint encoding here. Logical values will
-      // usually be small.
-      EncodeUint32(&s, uint32_t(k.logical));
-    }
+    EncodeTimestamp(s, k.wall_time, k.logical);
   }
   s.push_back(char(s.size() - k.key.len));
   return s;
@@ -1038,13 +1042,13 @@ class BaseDeltaIterator : public rocksdb::Iterator {
  public:
   BaseDeltaIterator(rocksdb::Iterator* base_iterator,
                     rocksdb::WBWIIterator* delta_iterator,
-                    bool prefix)
+                    bool prefix_same_as_start)
       : current_at_base_(true),
         equal_keys_(false),
         status_(rocksdb::Status::OK()),
         base_iterator_(base_iterator),
         delta_iterator_(delta_iterator),
-        prefix_same_as_start_(prefix) {
+        prefix_same_as_start_(prefix_same_as_start) {
     merged_.data = NULL;
   }
 
@@ -1715,11 +1719,13 @@ DBStatus DBMerge(DBEngine* db, DBKey key, DBSlice value) {
 }
 
 DBStatus DBImpl::Get(DBKey key, DBString* value) {
+  rocksdb::ReadOptions read_opts;
   DBGetter base(rep, read_opts, EncodeKey(key));
   return base.Get(value);
 }
 
 DBStatus DBBatch::Get(DBKey key, DBString* value) {
+  rocksdb::ReadOptions read_opts;
   DBGetter base(rep, read_opts, EncodeKey(key));
   if (updates == 0) {
     return base.Get(value);
@@ -1734,6 +1740,8 @@ DBStatus DBWriteOnlyBatch::Get(DBKey key, DBString* value) {
 }
 
 DBStatus DBSnapshot::Get(DBKey key, DBString* value) {
+  rocksdb::ReadOptions read_opts;
+  read_opts.snapshot = snapshot;
   DBGetter base(rep, read_opts, EncodeKey(key));
   return base.Get(value);
 }
@@ -1920,36 +1928,28 @@ DBEngine* DBNewBatch(DBEngine *db, bool writeOnly) {
   return new DBBatch(db);
 }
 
-DBIterator* DBImpl::NewIter(bool prefix) {
+DBIterator* DBImpl::NewIter(rocksdb::ReadOptions* read_opts) {
   DBIterator* iter = new DBIterator;
-  rocksdb::ReadOptions opts = read_opts;
-  opts.prefix_same_as_start = prefix;
-  opts.total_order_seek = !prefix;
-  iter->rep.reset(rep->NewIterator(opts));
+  iter->rep.reset(rep->NewIterator(*read_opts));
   return iter;
 }
 
-DBIterator* DBBatch::NewIter(bool prefix) {
+DBIterator* DBBatch::NewIter(rocksdb::ReadOptions* read_opts) {
   DBIterator* iter = new DBIterator;
-  rocksdb::ReadOptions opts = read_opts;
-  opts.prefix_same_as_start = prefix;
-  opts.total_order_seek = !prefix;
-  rocksdb::Iterator* base = rep->NewIterator(opts);
+  rocksdb::Iterator* base = rep->NewIterator(*read_opts);
   rocksdb::WBWIIterator* delta = batch.NewIterator();
-  iter->rep.reset(new BaseDeltaIterator(base, delta, prefix));
+  iter->rep.reset(new BaseDeltaIterator(base, delta, read_opts->prefix_same_as_start));
   return iter;
 }
 
-DBIterator* DBWriteOnlyBatch::NewIter(bool prefix) {
+DBIterator* DBWriteOnlyBatch::NewIter(rocksdb::ReadOptions* read_opts) {
   return NULL;
 }
 
-DBIterator* DBSnapshot::NewIter(bool prefix) {
+DBIterator* DBSnapshot::NewIter(rocksdb::ReadOptions* read_opts) {
+  read_opts->snapshot = snapshot;
   DBIterator* iter = new DBIterator;
-  rocksdb::ReadOptions opts = read_opts;
-  opts.prefix_same_as_start = prefix;
-  opts.total_order_seek = !prefix;
-  iter->rep.reset(rep->NewIterator(opts));
+  iter->rep.reset(rep->NewIterator(*read_opts));
   return iter;
 }
 
@@ -1995,7 +1995,32 @@ DBStatus DBSnapshot::GetStats(DBStatsResult* stats) {
 }
 
 DBIterator* DBNewIter(DBEngine* db, bool prefix) {
-  return db->NewIter(prefix);
+  rocksdb::ReadOptions opts;
+  opts.prefix_same_as_start = prefix;
+  opts.total_order_seek = !prefix;
+  return db->NewIter(&opts);
+}
+
+DBIterator* DBNewTimeBoundIter(DBEngine* db, DBTimestamp min_ts, DBTimestamp max_ts) {
+  const std::string min = EncodeTimestamp(min_ts);
+  const std::string max = EncodeTimestamp(max_ts);
+  rocksdb::ReadOptions opts;
+  opts.total_order_seek = true;
+  opts.table_filter = [min, max](const rocksdb::TableProperties& props) {
+    auto userprops = props.user_collected_properties;
+    auto tbl_min = userprops.find("crdb.ts.min");
+    if (tbl_min == userprops.end() || tbl_min->second.empty()) {
+      return true;
+    }
+    auto tbl_max = userprops.find("crdb.ts.max");
+    if (tbl_max == userprops.end() || tbl_max->second.empty()) {
+      return true;
+    }
+    // If the timestamp range of the table overlaps with the timestamp range we
+    // want to iterate, the table might contain timestamps we care about.
+    return max.compare(tbl_min->second) >= 0 && min.compare(tbl_max->second) <= 0;
+  };
+  return db->NewIter(&opts);
 }
 
 void DBIterDestroy(DBIterator* iter) {