--- a/pkg/testutils/files.go
+++ b/pkg/testutils/files.go
@@ -1,0 +1,42 @@
+// Copyright 2017 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+//
+// Author: Nikhil Benesch (nikhil.benesch@gmail.com)
+
+package testutils
+
+import (
+	"bufio"
+	"io"
+	"io/ioutil"
+	"os"
+	"path/filepath"
+)
+
+// ReadAllFiles reads all of the files matching pattern, thus ensuring they are
+// in the OS buffer cache.
+func ReadAllFiles(pattern string) {
+	matches, err := filepath.Glob(pattern)
+	if err != nil {
+		return
+	}
+	for _, m := range matches {
+		f, err := os.Open(m)
+		if err != nil {
+			continue
+		}
+		_, _ = io.Copy(ioutil.Discard, bufio.NewReader(f))
+		f.Close()
+	}
+}