See commit messages for details. Benchmarking results:

```
BenchmarkClusterEmptyIncrementalBackup
       2	 642326712 ns/op	  19.21 MB/s

COCKROACH_TIME_BOUND_ITERATORS=1 BenchmarkClusterEmptyIncrementalBackup
      10	 123180991 ns/op	 100.16 MB/s

BenchmarkTimeBoundIterate/LoadFactor1.00NormalIterator-8         	       2	 692584644 ns/op	  73.93 MB/s
BenchmarkTimeBoundIterate/LoadFactor1.00TimeBoundIterator-8      	       2	 726511294 ns/op	  70.47 MB/s
BenchmarkTimeBoundIterate/LoadFactor0.50NormalIterator-8         	       2	 723423248 ns/op	  70.77 MB/s
BenchmarkTimeBoundIterate/LoadFactor0.50TimeBoundIterator-8      	       3	 361400000 ns/op	 141.67 MB/s
BenchmarkTimeBoundIterate/LoadFactor0.10NormalIterator-8         	       2	 682549781 ns/op	  75.01 MB/s
BenchmarkTimeBoundIterate/LoadFactor0.10TimeBoundIterator-8      	      20	  55027403 ns/op	 930.45 MB/s
BenchmarkTimeBoundIterate/LoadFactor0.05NormalIterator-8         	       2	 730835631 ns/op	  70.06 MB/s
BenchmarkTimeBoundIterate/LoadFactor0.05TimeBoundIterator-8      	      50	  24495108 ns/op	2090.21 MB/s
BenchmarkTimeBoundIterate/LoadFactor0.00NormalIterator-8         	       2	 698786270 ns/op	  73.27 MB/s
BenchmarkTimeBoundIterate/LoadFactor0.00TimeBoundIterator-8      	    1000	   1275639 ns/op	40136.73 MB/s
PASS
ok  	github.com/cockroachdb/cockroach/pkg/ccl/storageccl/engineccl	20.821s
```

`BenchmarkTimeBoundIterate` is incredibly unrealistic as the load factor drops; `BenchmarkClusterEmptyIncrementalBackup` essentially has a load factor of 0.0 as well.

We'll probably want to remove the `COCKROACH_TIME_BOUND_ITERATORS` before merging, but figured I'd leave it in for now. Also need to properly land the `c-rocksdb` patch.