@danhhz 2017-04-17
This is great! I took an initial pass and the general structure looks good. A detailed review will probably have to wait till tomorrow. A few initial comments:
- Can you run the benchmarks with -count=5 and use benchstat to get the +- numbers
- The name for table_properties_collector is going to need some bikeshedding (sst_filter?)
- We'll probably also want some correctness tests comparing the result of the normal iterator and the time bound one

Adding @petermattis as a reviewer too