@bdarnell 2017-04-17
LGTM. What's the rocksdb patch? And note that #14840 changes the mechanism for rocksdb patches.

---

Reviewed 7 of 8 files at r1, 5 of 5 files at r2, 4 of 4 files at r3.
Review status: all files reviewed at latest revision, 2 unresolved discussions, some commit checks failed.

---

*[pkg/ccl/storageccl/engineccl/bench_test.go, line 44 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-KhxxmN6jgYEfCBhDZOa:-KhxxmN6jgYEfCBhDZOb:bdlnnr) ([raw file](https://github.com/cockroachdb/cockroach/blob/3a5639a9fbcd2863f8d334dbe1f0281ad8311f49/pkg/ccl/storageccl/engineccl/bench_test.go#L44)):*
> ```Go
> // be created:
> //
> // key-1#2  ┐
> ```

Nice use of ASCII^H^H^H^H Unicode art

---

*[pkg/storage/engine/rocksdb.go, line 1131 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-Khxz7Xja0VNlmdNKXzV:-Khxz7Xja0VNlmdNKXzW:b-5mzkec) ([raw file](https://github.com/cockroachdb/cockroach/blob/3a5639a9fbcd2863f8d334dbe1f0281ad8311f49/pkg/storage/engine/rocksdb.go#L1131)):*
> ```Go
> func (r *rocksDBBatch) NewTimeBoundIterator(start, end hlc.Timestamp) Iterator {
> 	panic("not implemented")
> 	if r.writeOnly {
> ```

Why include this body in an unimplemented method? Do we expect to implement this soon?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-:-KhxvtOHVbegE5vutAmV:bqi554j)*
<!-- Sent from Reviewable.io -->