@benesch 2017-04-17
RocksDB patch is here: https://github.com/cockroachdb/c-rocksdb/blob/c31d6778babd00fecee0e469336b250f3d8f1613/table_properties_inspector.patch

@danhhz re the patch, we may want to submit upstream for naming advice too.

---

Review status: 7 of 16 files reviewed at latest revision, 2 unresolved discussions.

---

*[pkg/ccl/storageccl/engineccl/bench_test.go, line 44 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-KhxxmN6jgYEfCBhDZOa:-Khy2ai8GjfcUBJbsfFB:b32vhw) ([raw file](https://github.com/cockroachdb/cockroach/blob/3a5639a9fbcd2863f8d334dbe1f0281ad8311f49/pkg/ccl/storageccl/engineccl/bench_test.go#L44)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary><blockquote>

Nice use of ASCII^H^H^H^H Unicode art
</blockquote></details>

:)

---

*[pkg/storage/engine/rocksdb.go, line 1131 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-Khxz7Xja0VNlmdNKXzV:-Khy2pQaLq6Te1nRsFOa:beoe14g) ([raw file](https://github.com/cockroachdb/cockroach/blob/3a5639a9fbcd2863f8d334dbe1f0281ad8311f49/pkg/storage/engine/rocksdb.go#L1131)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary><blockquote>

Why include this body in an unimplemented method? Do we expect to implement this soon?
</blockquote></details>

Nope, just a rebase gone wrong. Fixed.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-:-Khy33iH7s82tLiI5R56:blms5l9)*
<!-- Sent from Reviewable.io -->