@petermattis 2017-04-18
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> modulo comment about how `ReadOptions` is being passed. The RocksDB patch is remarkably small.

---

Review status: 7 of 16 files reviewed at latest revision, 5 unresolved discussions, some commit checks failed.

---

*[pkg/storage/engine/db.cc, line 206 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-KhyOQXN9g5vyjYJhUai:-KhyOQXO6EMTWYpc6Iws:b9twuv2) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/storage/engine/db.cc#L206)):*
> ```cpp
>   if (logical != 0) {
>     // TODO(peter): Use varint encoding here. Logical values will
>     // usually be small.
> ```

I think this ship has sailed. Should probably remove this TODO.

---

*[pkg/storage/engine/db.cc, line 1935 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-KhyOlIpvrUbnV-a6pNY:-KhyOlIq-pccopOXW43Y:bfkeaq) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/storage/engine/db.cc#L1935)):*
> ```cpp
> }
> 
> DBIterator* DBImpl::NewIter(rocksdb::ReadOptions* read_opts) {
> ```

Can't you pass `const rocksdb::ReadOptions&` here? My C++ is rusty, but I think the `table_properties_inspector` will be copied correctly. I'm pretty sure `read_options` get passed around by const reference in the RocksDB code, so if it doesn't work here we'd need to make `table_properties_inspector` an `auto_ptr` or something like that.

---

*[pkg/storage/engine/db.cc, line 2012 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-KhyPG-yVTx7JpUneR9I:-KhyPG-yVTx7JpUneR9J:b-3ppodk) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/storage/engine/db.cc#L2012)):*
> ```cpp
> DBIterator* DBNewTimeBoundIter(DBEngine* db, DBTimestamp min_ts, DBTimestamp max_ts) {
>   std::string min = EncodeTimestamp(min_ts);
>   std::string max = EncodeTimestamp(max_ts);
> ```

Doesn't really matter, but `min` and `max` can both be `const std::string`.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-:-KhyPcRpjav_buPiMCdQ:bvht44w)*
<!-- Sent from Reviewable.io -->