@benesch 2017-04-20
benchstat results:

```
name                                    old time/op    new time/op     delta
ClusterEmptyIncrementalBackup    68.5ms ± 7%     44.8ms ±33%   -34.68%  (p=0.008 n=5+5)

name                                    old speed      new speed       delta
ClusterEmptyIncrementalBackup  3.61GB/s ± 7%  10.81GB/s ±47%  +199.64%  (p=0.008 n=5+5)
```

```
name                             old time/op   new time/op      delta
TimeBoundIterate/LoadFactor1.00   62.3ms ± 2%      64.7ms ± 3%      +3.75%  (p=0.016 n=5+5)
TimeBoundIterate/LoadFactor0.50   63.2ms ± 4%      35.8ms ± 4%     -43.29%  (p=0.008 n=5+5)
TimeBoundIterate/LoadFactor0.10   64.9ms ± 5%       5.4ms ± 7%     -91.66%  (p=0.008 n=5+5)
TimeBoundIterate/LoadFactor0.05   67.4ms ± 7%       2.5ms ± 7%     -96.36%  (p=0.008 n=5+5)
TimeBoundIterate/LoadFactor0.00   66.1ms ± 2%       0.2ms ± 6%     -99.69%  (p=0.008 n=5+5)

name                             old speed     new speed        delta
TimeBoundIterate/LoadFactor1.00  822MB/s ± 2%     792MB/s ± 3%      -3.59%  (p=0.016 n=5+5)
TimeBoundIterate/LoadFactor0.50  811MB/s ± 4%    1430MB/s ± 4%     +76.26%  (p=0.008 n=5+5)
TimeBoundIterate/LoadFactor0.10  790MB/s ± 5%    9477MB/s ± 7%   +1099.85%  (p=0.008 n=5+5)
TimeBoundIterate/LoadFactor0.05  761MB/s ± 7%   20930MB/s ± 7%   +2651.66%  (p=0.008 n=5+5)
TimeBoundIterate/LoadFactor0.00  774MB/s ± 2%  254069MB/s ± 6%  +32709.30%  (p=0.008 n=5+5)
```

---

Review status: 2 of 19 files reviewed at latest revision, 13 unresolved discussions, some commit checks failed.

---

*[pkg/ccl/storageccl/engineccl/bench_test.go, line 40 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-Ki6eJ9eGNFkCc1_43r4:-Ki6qG_7cvCDsAsymxsQ:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/ccl/storageccl/engineccl/bench_test.go#L40)):*
<details><summary><i>Previously, danhhz (Daniel Harrison) wrote…</i></summary><blockquote>

as discussed offline, the timestamp randomization is misleading if it's not meaningful
</blockquote></details>

Done.

---

*[pkg/ccl/storageccl/engineccl/bench_test.go, line 175 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-Ki6dvy60WCBq_CpAExi:-Ki6oNUxWa1i5voNH-wR:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/ccl/storageccl/engineccl/bench_test.go#L175)):*
<details><summary><i>Previously, danhhz (Daniel Harrison) wrote…</i></summary><blockquote>

I think the `a` just adds confusion, it's fine to compare to `n` and use it in the Fatalf
</blockquote></details>

Done.

---

*[pkg/ccl/storageccl/engineccl/bench_test.go, line 183 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-Ki6ezVDvbn5klBpeT4g:-KiB2A6pFuKh1H9WdMI0:b-kwn0c3) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/ccl/storageccl/engineccl/bench_test.go#L183)):*
<details><summary><i>Previously, danhhz (Daniel Harrison) wrote…</i></summary><blockquote>

We should add some correctness tests, too
</blockquote></details>

Still WIP.

---

*[pkg/ccl/storageccl/engineccl/bench_test.go, line 185 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-Ki6dXdY67W8BqPyWbz9:-Ki6olv7SeMGmgilexuB:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/ccl/storageccl/engineccl/bench_test.go#L185)):*
<details><summary><i>Previously, danhhz (Daniel Harrison) wrote…</i></summary><blockquote>

maybe make two levels of sub-benchmark. the first for load factor and the second for which iterator
</blockquote></details>

Done.

---

*[pkg/ccl/storageccl/engineccl/mvcc.go, line 63 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-Ki6btJY0TS47p2eE1ac:-Ki6tFnPWa_HIKGaBSfs:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/ccl/storageccl/engineccl/mvcc.go#L63)):*
<details><summary><i>Previously, danhhz (Daniel Harrison) wrote…</i></summary><blockquote>

Having a Reset was nice, maybe make this take just the time bounds and leave the key bounds back where they were?
</blockquote></details>

Done.

---

*[pkg/storage/engine/db.cc, line 206 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-KhyOQXN9g5vyjYJhUai:-Ki6p8o7XP-imtg8gsfN:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/storage/engine/db.cc#L206)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

I think this ship has sailed. Should probably remove this TODO.
</blockquote></details>

Done.

---

*[pkg/storage/engine/db.cc, line 1935 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-KhyOlIpvrUbnV-a6pNY:-Khz85N9QJmsWANjevkM:bwm4j32) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/storage/engine/db.cc#L1935)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

Can't you pass `const rocksdb::ReadOptions&` here? My C++ is rusty, but I think the `table_properties_inspector` will be copied correctly. I'm pretty sure `read_options` get passed around by const reference in the RocksDB code, so if it doesn't work here we'd need to make `table_properties_inspector` an `auto_ptr` or something like that.
</blockquote></details>

You're right, of course. Much cleaner this way, too. Done.

---

*[pkg/storage/engine/db.cc, line 2012 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-KhyPG-yVTx7JpUneR9I:-Ki6pD8gIWHSUT1dHtrC:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/storage/engine/db.cc#L2012)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

Doesn't really matter, but `min` and `max` can both be `const std::string`.
</blockquote></details>

Done.

---

*[pkg/storage/engine/engine.go, line 127 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-Ki6cpXoVie2mu7m0V2i:-Ki6tJ2eI08-0hKzcK2T:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/storage/engine/engine.go#L127)):*
<details><summary><i>Previously, danhhz (Daniel Harrison) wrote…</i></summary><blockquote>

time range
</blockquote></details>

Done.

---

*[pkg/storage/engine/rocksdb.go, line 1313 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-Ki6ct69PHLSNEjkBww0:-Ki6u5U3v177ZYL4qwBU:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/storage/engine/rocksdb.go#L1313)):*
<details><summary><i>Previously, danhhz (Daniel Harrison) wrote…</i></summary><blockquote>

ditto "time range"
</blockquote></details>

Done.

---

*[pkg/storage/engine/rocksdb.go, line 1318 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-Ki6dHZn_o8JQzReUk3Y:-Ki6u5nye04bR5sPq5rP:b-q0ppy8) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c9529d97d0d8f1bf02fe801f22a4e41b0078046/pkg/storage/engine/rocksdb.go#L1318)):*
<details><summary><i>Previously, danhhz (Daniel Harrison) wrote…</i></summary><blockquote>

Is it safe to share the same pool as the normal iterators?
</blockquote></details>

Ha! Absolutely not. Fixed.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/15004#-:-KiB8g9jPEAS6G3FJWj7:b-b482h0)*
<!-- Sent from Reviewable.io -->